refactor: Modularize streaming functionality with dedicated handlers and validation
